+++ CircuitSimulator/src/Components/Voltmeter.java
@@ -3,7 +3,7 @@ package Components;
 
 public class Voltmeter extends Component {
 	
-	private int readVoltage;
+	private double readVoltage;
 	private final static int MAX_ROTATION = 2;
 
 	public Voltmeter(int x, int y, int type) {
@@ -13,11 +13,11 @@ public class Voltmeter extends Component {
 		maxRotation = MAX_ROTATION;
 	}
 
-	public int getReadVoltage() {
+	public double getReadVoltage() {
 		return readVoltage;
 	}
 
-	public void setReadVoltage(int readVoltage) {
+	public void setReadVoltage(double readVoltage) {
 		this.readVoltage = readVoltage;
 	}
